agentcore-runtime-agent/agent_pdz_02/agent/requirements.txt
fastapi
uvicorn
pydantic
mcp
httpx
strands-agents
boto3
requests
awscli
mcp-proxy-for-aws
